Added a cast to double to correct compilation problem, caused by Olivier's change of this file.  Apparently this change passed all the tests at Lisa, but not at ApSTAT.  Probably due to library version differences in std::log, which at ApSTAT leads it to cinsider std::log(int) ambiguous (hestating between float and double) but not at Lisa.

+++ plearn/math/TMat_maths_impl.h
@@ -5627,7 +5627,7 @@ T det(const TMat<T>& LU, int detsign, bool log_det = false)
             minus = !minus;
             detsign = - detsign;
         }
-        determinant = pl_log(detsign);
+        determinant = pl_log(double(detsign));
     }
     int mod = LU.mod();
     int n = LU.width();
